refactor: share component validation
